Spec helpers can use a list of events
Helpers `given_events` and `then_events` accept a varargs events for
convenience to pass in a few events. Sharing a list of events accross
specs however, required splatting the event list.

With this patch the events list is flattened by 1 so that either varargs
or a list can be used.