Order events for replay by stream and sequence
Event migrations can cause events to be inserted in existing event
streams. When replaying events by their event ID causes these events to
be replayed at the end of the stream.

This patch changes the order of events for replay to the ID of the
stream and within a stream the event's sequence number. In most
situations the result should be identical. In cases where not, the
sequence number ordering is what you'd expect (since this is also how
Sequent loads an aggregate from the event store).