Add dry run option
When using dry run the given commands and resulting events are not stored
but recorded for inspection purposes.
It will also record which Projectors and Workflows would have been invoked
as a result of the Events applied.
Due to the fact that the Sequent.configuration is a singleton and shared
among Threads dry run is not Thread safe.