Split creating indexes from creating tables
Doing a full replay of a large projector can take a lot of time.
This can be greatly reduced by creating the indexes after
the table has been filled.
This works as follows: Create a file called table_name.indexes.sql in the same directory as the table_name.sql. Sequent will then pick this up and run after the `migrate_online` ran, but before the `migrate_offline` phase.

Pre-cautions:
- Don't put any UNIQUE INDEXES in this file because then it will only fail after the `migrate_online` is done. Both scenario's will fail of course but this is just sooner.
- If you don't use the `ReplayOptimizedPostgresPersistor` it might
actually slow you down because of the intermediate queries that will
happen in that case.