Add support for multiple database applications
In ActiveRecord 6+ multiple databases can be used under specific roles. For example a read-replica can be used for read only roles. This requires a different database configuration structure and connection handling.

This PR allows Sequent to connect to databases configured in this way, though due to the nature of event sourcing it can only connect to a single database, which should always be the primary database.