Provide a proper way of running Rails migrations next to Sequent migrations
The problem occurs when rails migrations are run after the view_schema and sequent_schema
are created and the migrations run when `schema_search_path` is set to contain both
the active record schema as well and the sequent schema's.
To get around this one must wrap `rails db:migrate` with a custom task in order to
set the `schema_search_path` to only contain the active record schema.
A better way would be to prepend something to `Rake::Task['db:migrate']` but there is only
`enchance` which will add a task as last one of the prerequisites.