Replay events using BETWEEN predicate on aggregate_id
Instead of using a prefix with a separate index. This will allow us to
drop the `replay_events` index on `event_type,
substring(aggregate_id::text, 1, 3)` since it uses the existing
`unique_event_per_aggregate` index instead.

This implementation is slower though when only a few event types need
to be replayed, instead of all events. Not sure if this is a big
problem or not.