Ensure every record class has an index on aggregate_id as documented
The implementation did not match the documentation as no index on
`aggregate_id` was created, even when the record class supported this
attribute. The index was also not added when indexes where specified
explicitly.

Now the `aggregate_id` index is always added when supported and for
record classes where no indexes are defined it will be the only
index (instead of no indexes).