Use meta-programming instead of eval to generate structs
The in-memory structs can easily be generated using "ordinary"
meta-programming. The only difference is that the struct classes are
anonymous since there is no longer a constant referring to the
class. This also avoids polluting the Ruby namespace.

Also add the missing `eql?` override to ensure the records are stored
correctly in the record `Set`.