Handle existing sequent schema when using active-record 7.1
The active record 7.1 schema dumper will include all non-`public`
schemas. This breaks the sequent table creation since it expect the
schema to not exist at all. So now we additionally check for the
presence of the `event_records` table or view to determine if the
schema file needs to be loaded.