Use separate snapshot table and composite primary keys
Due to the use of `aggregate_id, sequence_number` as the primary key
the snapshots can no longer be stored in the event records table, so
they now have their own table. The event store API has been extended
to allow storing and loading snapshots independently, but snapshots
are also returned when loading the events for an aggregate, like
before.

Since snapshots are now in their own table it is no longer needed to
associate a command with a snapshot, but for now the `TakeSnapshot`
command is still present.