Allow passing the same stream multiple times
Some tests rely on a specific ordering of storing events so we cannot
just group these events by aggregate and pass them to the event
store. Instead, each event is associated with a stream and the entire
set is passed to the event store. So the event store needs to make
sure `upsert_all` does not get the same `aggregate_id` multiple times
when upserting the stream records.