Optimize replay of events by scanning partitioned events directly
This avoids using the `event_records` view which joins with the
`aggregates` table, which is unnecessary in this case.

Also add "before" and "after" groups to ensure we capture all possible
events, in case statistical scanning is used to count the number of
events to determine the event replay groups.