Track snapshot high watermark so that snapshotting can be prioritized
After deleting all snapshots the first aggregates that we want to
restore are the ones with the highest number of events. We track the
high-water mark of the stored snapshots per aggregate to know which
aggregates need to be snapshotted first.

In normal operation the xact_id is used to prioritize snapshotting
ordering.