Protect against concurrency problems at read committed isolation level
With read committed isolation level it is possible for the
`load_events` function to fail when another session updates the
aggregate's `events_partition_key`, since reads are not repeatable at
this isolation level.

By using a single query to load an aggregate, its snapshot, and its
event's all queries use the same database snapshot ensuring a
consistent set of data.

When storing the events order lock the aggregate row and inserts in
`aggregate_id` order to concurrency issues (deadlocks or foreign-key
violations).