Limit the number of aggregates that are scheduled for snapshotting
When selecting aggregates for snapshotting (aggregates with an
outdated snapshot) there can never be more than the specified limit
aggregates returned that are currently not yet scheduled for
snapshotting (snapshot_scheduled_at is null).

This allows background tasks to finish snapshotting aggregates before
new (or the same) aggregates are scheduled for snapshotting again.