Avoid exhausting type ids due to transaction rollbacks
Type ids (for aggregate, command, and event types) use the `SMALLINT`
data type to optimize storage usage. However, when a sequence is used
to generated ids the ids can be dropped on transaction rollback. This
can cause the ids to be quickly exhausted when something is
misconfigured.

This change uses the highest id + 1 for the next id. This is slow and
requires table locking, but since new types are rarely added this
should not impact normal performance.