Track event partition key changes for aggregates
Before the partition key of an aggregate and its events where
immediately updated. Unfortunately the view schema migration relies on
a stable ordering of aggregates based on `events_partition_key,
aggregate_id`. But if this changes while a migration is running the
aggregate may not be replayed at all or be replayed multiple times.

With this change the partition key updates are only logged when an
aggregate is updated and a separate maintenance task is used to
perform the actual updates. This maintenance task can only be run when
there is no view schema migration active.