Support storing multiple snapshot versions in the event store
Pass snapshot version information when retrieving and storing
snapshots from/to the database.

A hash of aggregate types to supported snapshot version is passed on
every call that needs it so the snapshot can be correctly
selected. The default implementation queries all descendants of
AggregateRoot for the supported snapshot version. This can be
configured with the `aggregate_snapshot_versions` setting.