Make projector state activation optional to ease upgrading and testing
As long as the projector_states table is empty it is assumed the
current Sequent configuration is correct. This allows easier unit
testing (since there is no need to activate a configuration every time
it changes in the test) and allows for backwards compatibility with
the current deployment process (using online and offline migration).