Do not use the `payload` method in `then_events` test helper
The `payload` method might have been overridden by event subclasses
which can cause the test to compare the wrong values. Instead
transform the event to JSON (as this format is saved to the database)
and remove the metadata fields for test comparisons.

In the future we want to ensure attributes do not override existing
methods to prevent this problem from occurring.