Support per-projector versioning to allow seamless upgrades
During an upgrade of an application most projectors will be the same,
so using a single global version will block many projectors that would
still work when using a view schema migration.

When using just background replay of projectors it is even possible to
keep the version of an updated projector the same if the projector is
back- and forward-compatible. This allows rolling upgrades without any
service operation.