Set lock timeout to avoid deadlocks when completing the replay
When completing the replay an exclusive lock needs to be acquired on
the tables in the view schema, which may be locked by the live systems
due to queries. Since the lock ordering cannot be guaranteed in this
case a deadlock may result, and the PostgreSQL deadlock detector will
kill transactions, which may cause service interruption on the live
system.

In this case it is preferrable that the replay completion process
fails as it can simply be retried until it succeeds.