Scan a subset of the events table to determine replay groups
When the events table is large we only need to sample part of the data
to determine the replay group boundaries. PostgreSQL can do this for
us using `TABLESAMPLE` where you can specify a percentage of the data
that needs to be scanned.

Then we use the PostgreSQL `ntile` window function to split the
matching events into boundaries based on partition key and aggregate
id.