Only use tablesample if number of estimated events is high
With tablesampling it is impossible to use an index and when the
number of estimatedxpected events is low we may not find any events,
so we cannot determine the groups. So in this case it is better to use
the indexes to find the matching events and use this as input for
grouping.