Avoid exclusive lock when there are no unknown active projectors
First query for the existence of unknown projectors before trying to
acquire an exclusive lock on the `projector_states` table.

On a busy system getting an exclusive lock can take some time and will
block projectors while waiting for the lock. Normally there are no
unknown active projectors so needing a lock should be rare.