Support replaying rojectors that use a databse view as managed table
In some cases it is useful to use a database view as a managed table
in the projector, for example when renaming a column or table while
needing to support the old and new version of a projector without
requiring downtime to perform the replay. In this case the tables used
by the view also need to be included in the replay schema.