Ensure archived tables do not have any dependent database objects
Archived tables should not be used anymore since they are replaced by
the replayed tables. However, if you use views, stored procedures,
foreign keys, etc that referenced the original view schema tables they
will now reference the archived tables, leading to surprising result.

PostgreSQL does not provide an easy query to find all these dependent
objects, so instead we start a sub-transaction (using a SAVEPOINT)
that tries to drop the archived tables without the CASCADE option. If
the drop succeeds only the sub-transaction is rolled back (so that the
tables are not actually dropped) and the activation is completed.

Otherwise the activation is aborted and an error is raised. You can
fix this by modifying the schema (re-creating a view, etc) using the
`projectors_replayer_after_activate_hook`.