Return events in creation order from event store when loading from mark
For testing the `then_events` helper retrieves and compares the events
since the last executed commands. This works fine for the fake event
store (where events are returned in creation order) but for the
PostgreSQL event store events might be returned in any order.

Now use the `created_at` field to return the events in creation order
for both the real and fake event store, breaking ties using aggregate
id and sequence number.