Notify callback when an aggregate is loaded that needs a new snapshot
This allows the user of Sequent to take a snapshot of the aggregate
and store it.

You can take the snapshot from the passed in aggregate but must store
the snapshot in the same transaction, otherwise non-committed events
might be part of the snapshot.

Another option is to reload the aggregate for snapshotting using a
separate transaction (for example, using a background job).