Overwrite existing snapshot when storing newer version
In case snapshots are created concurrently (due to notifications when
loading an aggregate) the last one created should overwrite any older
versions.

This may also help when a snapshot is corrupted and a newer snapshot
is created, although in this case it is better to delete the corrupted
snapshot first.